+++ client/src/pages/Episodes/Episodes.js
@@ -2,6 +2,7 @@ import React, {Component} from "react";
 
 import { EpisodeList, EpisodeListItem } from "../../components/EpisodeList";
 import Jumbotron from "../../components/JumbotronEps";
+import Navbar from "../../components/Navbar";
 import PropTypes from 'prop-types';
 
 import API from "../../utils/API";
@@ -36,6 +37,7 @@ class Episodes extends Component {
   render() {
     return (
     <div>
+      <Navbar />
       <Jumbotron
         name={this.state.feed.name}
         image={this.state.feed.image}
