+++ client/src/pages/Home/Home.js
@@ -1,0 +1,15 @@
+import React from 'react';
+
+import SearchAndResults from "../../components/SearchAndResults";
+import Navbar from "../../components/Navbar";
+
+const Home = () => {
+  return (
+    <div>
+      <Navbar />
+      <SearchAndResults />
+    </div>
+  );
+}
+
+export default Home;
