+++ client/src/pages/Login/Login.js
@@ -23,9 +23,8 @@ export default withAuth(class Login extends Component {
   }
 
   render() {
-    if (this.state.authenticated === null) return null;
     return this.state.authenticated ?
-      <Redirect to={{ pathname: '/' }}/> :
+      <Redirect to={{ pathname: '/profile' }}/> :
       <LoginForm baseUrl={this.props.baseUrl} />;
   }
 });
